Split up Redux files
