+++ assets/redux/actions/curationResultActions.js
@@ -1,4 +1,4 @@
-import api from "../api";
+import api from "../../api";
 
 export const actionTypes = {
   SET_RESULT: "SET_RESULT",
