+++ assets/redux/selectors/curationResultSelectors.js
