+++ assets/redux/store.js
@@ -1,7 +1,7 @@
 import { applyMiddleware, combineReducers, createStore } from "redux";
 import thunk from "redux-thunk";
 
-import { curationResultReducer } from "./reducers";
+import curationResultReducer from "./reducers/curationResultReducer";
 
 const rootReducer = combineReducers({
   curationResult: curationResultReducer,
